feat(PEER-729): use custom theme and refactor folder structure
