+++ apps/kernel/management-shell-browser/src/app.tsx
@@ -1,0 +1,10 @@
+import { AppProviders } from './app-providers';
+import { AppRouter } from './app-router';
+
+export const App = () => {
+  return (
+    <AppProviders>
+      <AppRouter />
+    </AppProviders>
+  );
+};
