+++ apps/kernel/management-shell-browser/src/domains/errors/components/error-boundary.tsx
