+++ apps/kernel/management-shell-browser/src/domains/feature-flags/components/flagged-component.types.ts
@@ -1,0 +1,5 @@
+export interface FlaggedComponentProps {
+  flagName: string;
+  children: React.ReactNode;
+  fallback?: React.ReactNode;
+}
