+++ apps/kernel/management-shell-browser/src/domains/feature-flags/components/if-flag-enabled.types.ts
@@ -1,0 +1,5 @@
+export interface IfFlagEnabledProps {
+  flagName: string;
+  children: React.ReactNode;
+  fallback?: React.ReactNode;
+}
