+++ apps/kernel/management-shell-browser/src/domains/feature-flags/flag-manager.ts
