+++ apps/kernel/management-shell-browser/src/domains/shared/utils/obj-from-array.ts
