+++ apps/kernel/management-shell-browser/src/domains/shared/utils/string-to-slug.ts
