+++ apps/kernel/management-shell-browser/src/domains/shared/utils/wait.ts
