+++ apps/kernel/management-shell-browser/src/domains/theme-settings/components/custom-theme-provider.types.ts
@@ -2,5 +2,5 @@ import { IThemeSettingsRepository } from '../repository';
 
 export interface ICustomThemeProviderProps {
   children: React.ReactNode;
-  settingsRepository: IThemeSettingsRepository;
+  themeSettingsRepository: IThemeSettingsRepository;
 }
