+++ apps/kernel/management-shell-browser/src/redux-store.ts
@@ -1,0 +1,19 @@
+import { AnyAction, ThunkAction, configureStore } from '@reduxjs/toolkit';
+import { TypedUseSelectorHook, useDispatch as useReduxDispatch, useSelector as useReduxSelector } from 'react-redux';
+import { themeSettingsSlice } from './domains/theme-settings/adapters/redux-slice';
+
+export const store = configureStore({
+  reducer: {
+    themeSettings: themeSettingsSlice.reducer,
+  },
+});
+
+export type IRootState = ReturnType<typeof store.getState>;
+
+export type IAppDispatch = typeof store.dispatch;
+
+export type IAppThunk = ThunkAction<void, IRootState, unknown, AnyAction>;
+
+export const useAppSelector: TypedUseSelectorHook<IRootState> = useReduxSelector;
+
+export const useAppDispatch = () => useReduxDispatch<IAppDispatch>();
