--- apps/kernel/management-shell-browser/src/store/index.ts
@@ -1,5 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-
-export const store = configureStore({
-  reducer: {},
-});
